词条配置

+++ api/service/image.js
@@ -5,15 +5,28 @@ import { dbFindOne } from './dbtools';
 import { deleteFile } from './qiniu';
 
 function getFileKey (url) {
-  const key = url.match(/(\w)+(\.jpg|\.png|\.jpeg)/)[0];
-  return key;
+  if (url !== '' && url !== null && url !== undefined) {
+    const key = url.match(/(\w)+(\.jpg|\.png|\.jpeg|\.gif)/);
+    if (key !== null) {
+      return key[0];
+    } else {
+      return '';
+    }
+  } else {
+    return '';
+  }
 }
 
 async function checkAndDeleteImg (data) {
   if (_.isObject(data)) {
+    /** {origin, tiny, min} */
     await deleteImgObj(data);
-  } else {
+  } else if (_.isString(data)) {
     await deleteUrl(data);
+    /** [{old, new}] */
+  } else if (_.isArray(data)) {
+    /** url */
+    await deleteArray(data);
   }
 }
 
@@ -38,15 +51,36 @@ async function deleteImgObj (imgObj) {
 }
 
 async function deleteUrl (url) {
-  const judge = await dbFindOne('Images', {
-    where: { $or: [
-      { origin: url },
-      { min: url },
-      { tiny: url }
-    ] }
-  });
-  if (judge === null) {
-    await deleteFile(getFileKey(url));
+  if (url !== '') {
+    const judge = await dbFindOne('Images', {
+      where: { $or: [
+        { origin: url },
+        { min: url },
+        { tiny: url }
+      ] }
+    });
+    if (judge === null) {
+      await deleteFile(getFileKey(url));
+    }
+  }
+}
+
+async function deleteArray (array) {
+  for (let item of array) {
+    const oldFileKey = getFileKey(item.oldUrl);
+    const newFileKey = getFileKey(item.newUrl);
+    if (oldFileKey !== newFileKey && oldFileKey !== '') {
+      const judge = await dbFindOne('Images', {
+        where: { $or: [
+          { origin: item.oldUrl },
+          { min: item.oldUrl },
+          { tiny: item.oldUrl }
+        ] }
+      });
+      if (judge === null) {
+        await deleteFile(getFileKey(item.oldUrl));
+      }
+    }
   }
 }
